LPD-62471 Apply to all publication FDS dateTime fields
