+++ modules/apps/change-tracking/change-tracking-web/src/main/java/com/liferay/change/tracking/web/internal/frontend/data/set/view/table/ScheduledPublicationsTableFDSView.java
@@ -1,0 +1,59 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.change.tracking.web.internal.frontend.data.set.view.table;
+
+import com.liferay.change.tracking.web.internal.constants.PublicationsFDSNames;
+import com.liferay.frontend.data.set.view.FDSView;
+import com.liferay.frontend.data.set.view.table.FDSTableSchema;
+import com.liferay.frontend.data.set.view.table.FDSTableSchemaBuilder;
+import com.liferay.frontend.data.set.view.table.FDSTableSchemaBuilderFactory;
+
+import java.util.Locale;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author David Truong
+ */
+@Component(
+	property = "frontend.data.set.name=" + PublicationsFDSNames.PUBLICATIONS_SCHEDULED,
+	service = FDSView.class
+)
+public class ScheduledPublicationsTableFDSView
+	extends BasePublicationsTableFDSView {
+
+	@Override
+	public FDSTableSchema getFDSTableSchema(Locale locale) {
+		FDSTableSchemaBuilder fdsTableSchemaBuilder =
+			_fdsTableSchemaBuilderFactory.create();
+
+		return fdsTableSchemaBuilder.add(
+			"name", "publication",
+			fdsTableSchemaField -> fdsTableSchemaField.setActionId(
+				"review-changes"
+			).setContentRenderer(
+				"actionLink"
+			).setSortable(
+				true
+			)
+		).add(
+			"description", "description"
+		).add(
+			addDateFDSTableSchemaField("dateScheduled", "publishing")
+		).add(
+			addDateFDSTableSchemaField("dateModified", "last-modified")
+		).add(
+			addDateFDSTableSchemaField("dateCreated", "create-date")
+		).add(
+			"ownerName", "owner"
+		).build();
+	}
+
+	@Reference
+	private FDSTableSchemaBuilderFactory _fdsTableSchemaBuilderFactory;
+
+}
